Wrote a node allocator.
Changed node->class to a bitfield instead of an enum to allow for
nodes that are both root and leaf (i.e. the first node in a tree).

Wrote a helper function to perform binary search through the keys
in a node and find the index of the child node containing a given
key.

Added high-level pesudocode for insert.