Fixed several bugs in radix_tree.c. Most of the low-level bit manipulation functions were missing some edge cases or invoking undefined behavior by shifting by more than the word width. radix_get_index is now extremely nasty but I can't find a clever way to avoid a branch that doesn't use gcc/clang builtins (oh how I wish those were portable :(   )
Removed an assert in alloc_node and fixed a stupid right below it.
Originally I was asserting that the tree did not have a root if a
parent node was not passed, but in the case that the tree has a single
node that is both root and leaf, when we go to decompress that path
by giving it a parent node, that node will be the new root (thus will
have no parent), but the tree *does* still have the old root.

Removed an assert in split_node_key that was redundant (the helper on
the next line asserts that anyway) ... somewhat of a sketchy change
but eh.

Fixed an inverted conditional in radix_tree_walk.