struct radix_node does not need a flag to denote leaf status. The prefix length is sufficient information to determine if a node is a leaf.
This commit thus removes the union containing the parent pointer and tag
and makes the setters/getters trivial. Will remove the setters and getters
and all calls to them in a later commit.

ammendment: void'ed arg to node_mark_leaf and removed const-ness of
parent pointer argument to set_parent to shut up the compiler