Debugging binomial heap. rekey() tests are now passing. (Turns out swapping a node with a child node is a lot more nuanced than I originally thought).
The peak() test is failing in some strange edge cases. Particularlly when
the minimum element of a heap happens to have a child node with the same
value. This looks to be a bug in pop and not peak, but its manifesting as a
double free in the peak test.