bloom: support union/intersection of filters.
This commit adds support for taking the union and intersection of two
bloom filters. See the doxygen comments for an explanation of their
behavior.

To accomplish this, this patch adds bloom_init_from. This allows one to
initialize a bloom filter with the same seeds as another filter. Without
this it is impossible to initialize two filters and have union or intersection
work because the two filters have different hash seeds. Also add a
bloom_same_class function that ensures that two filters are similar enough to
be able to take their union/intersection.

Finally, this patch adds tests for intersection and union and makes some
general improvements to the bloom filter tests.